requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
pandas==2.2.0
python-dotenv==1.0.1
